+++ DistributedSystemsWCF/DistributedSystemsHost/RicartAgrawala.cs
@@ -12,7 +12,7 @@ namespace DistributedSystems
         // --- Private Properties ----------------------------------------
         private LamportClock Clock;
         private bool Acquiring = false;
-        private long SentTimeStamp;
+        private uint SentTimeStamp;
         private bool IsLocked = false;
         private CountdownLatch Latch;
         private HashSet<string> Replied = new HashSet<string>();
@@ -115,11 +115,11 @@ namespace DistributedSystems
             IsLocked = true;
             Latch = null;
         }
-    
-    
-        public void MessageReceived(string ip, long k)
+
+
+        public void MessageReceived(string ip, uint k)
         {
-            Clock.EventReceive(new Tuple<long, string>(k, ip));
+            Clock.EventReceive(new Tuple<uint, string>(k, ip));
     
             // If not interested in critical section reply with 'OK'
             if (!NeedsToAccessCriticalSection)
@@ -131,7 +131,7 @@ namespace DistributedSystems
             else if (Acquiring)
             {
                 // I win (the result of this comparison should be the same on the sender as well
-                if (new LamportClock(SentTimeStamp, Node.Instance.Address).Compare(new Tuple<long, string>(k, ip)) < 0)
+                if (new LamportClock(SentTimeStamp, Node.Instance.Address).Compare(new Tuple<uint, string>(k, ip)) < 0)
                 {
                     PendingReplies.Enqueue(ip);
                 // I lose
